tap plugins: Replaced a lot of unnecessary, and potentially dangerous, unsigned integers with signed integers. This at least fixes a freeze in the reflector plugin.

+++ kokkinizita/tap-plugins-0.7.3/tap_eq.c
@@ -299,7 +299,7 @@ run_eq(LADSPA_Handle instance, unsigned long sample_count) {
 	biquad * filters = ptr->filters;
 	float fs = ptr->fs;
 
-	unsigned long pos;
+	int pos;
 	float samp;
 
 
@@ -415,7 +415,7 @@ run_adding_eq(LADSPA_Handle instance, unsigned long sample_count) {
 	biquad * filters = ptr->filters;
 	float fs = ptr->fs;
 
-	unsigned long pos;
+	int pos;
 	float samp;
