+++ kokkinizita/tap-plugins-0.7.3/tap_eqbw.c
@@ -361,7 +361,7 @@ run_eq(LADSPA_Handle instance, unsigned long sample_count) {
 	biquad * filters = ptr->filters;
 	float fs = ptr->fs;
 
-	unsigned long pos;
+	int pos;
 	float samp;
 
 
@@ -501,7 +501,7 @@ run_adding_eq(LADSPA_Handle instance, unsigned long sample_count) {
 	biquad * filters = ptr->filters;
 	float fs = ptr->fs;
 
-	unsigned long pos;
+	int pos;
 	float samp;
